add test for string enum
